+++ sdk/core/core-client-rest/src/sendRequest.ts
@@ -172,6 +172,11 @@ function getRequestBody(body?: unknown, contentType: string = ""): RequestBody {
     return { body: JSON.stringify(body) };
   }
 
+  if (firstType === "application/octet-stream" &&
+    body instanceof Uint8Array) {
+    return { body };
+  }
+
   if (ArrayBuffer.isView(body)) {
     if (body instanceof Uint8Array) {
       return { body: binaryArrayToString(body) };